fix bug

+++ src/main/java/com/sparky/lirenadmin/bo/SalesPerformanceBO.java
@@ -3,6 +3,7 @@ package com.sparky.lirenadmin.bo;
 import com.sparky.lirenadmin.entity.SalesPerformance;
 import com.sparky.lirenadmin.entity.ShopEmployee;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public interface SalesPerformanceBO {
 
     void reward(Long salesNo);
 
-    Integer sumSalePerformanceNum(ShopEmployee employee, Date today);
+    BigDecimal sumSalePerformanceNum(ShopEmployee employee, Date today);
 
     SalesPerformance getSalesPerf(Long id);
 }
