+++ src/main/java/com/sparky/lirenadmin/mapper/ext/SalesPerformanceMapperExt.java
@@ -3,12 +3,13 @@ package com.sparky.lirenadmin.mapper.ext;
 import com.sparky.lirenadmin.mapper.SalesPerformanceMapper;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Repository
 public interface SalesPerformanceMapperExt extends SalesPerformanceMapper {
 
-    int sumSalePerformanceNumByShop(Long shopNo, Date today);
+    BigDecimal sumSalePerformanceNumByShop(Long shopNo, Date today);
 
     /**
      * 查询当前员工和该员工下级的业绩数据
@@ -16,5 +17,5 @@ public interface SalesPerformanceMapperExt extends SalesPerformanceMapper {
      * @param today
      * @return
      */
-    int sumSalePerformanceNumByEmp(Long empNo, Date today);
+    BigDecimal sumSalePerformanceNumByEmp(Long empNo, Date today);
 }
